chore: better types
