+++ packages/contentful--create-contentful-app/package.json
@@ -13,7 +13,6 @@
     "node": ">=12",
     "npm": ">=6"
   },
-  "type": "module",
   "main": "lib/index.js",
   "lint-staged": {
     "*.js": [
@@ -63,6 +62,9 @@
   },
   "devDependencies": {
     "@types/degit": "^2.8.3",
-    "@types/validate-npm-package-name": "^3.0.3"
+    "@types/inquirer": "^8.2.0",
+    "@types/validate-npm-package-name": "^3.0.3",
+    "@typescript-eslint/eslint-plugin": "^5.11.0",
+    "@typescript-eslint/parser": "^5.11.0"
   }
 }
