+++ packages/contentful--create-contentful-app/src/types.d.ts
@@ -1,0 +1,3 @@
+declare module '@contentful/app-scripts' {
+  const createAppDefinition: { interactive: () => Promise<void> };
+}
